+++ src/audio.ml
@@ -21,6 +21,9 @@ let play_music music =
     Raylib.seek_music_stream music.t (Float.bound 0.1 music.loop_start.at Float.max_float);
   Raylib.update_music_stream music.t
 
+let play_continuous_sound state sound_name =
+  play_music (String.Map.find sound_name state.global.continuous_sounds)
+
 let get_area_music area_id area_musics = List.find (fun am -> List.mem area_id am.areas) area_musics
 let get_new_volume initial change = Float.bound 0. (initial +. change) 1.
 
@@ -42,17 +45,20 @@ let play_game_music game = play_music game.music.music
 let stop_music music = Raylib.stop_music_stream music
 let reset_music music = Raylib.seek_music_stream music.t 0.
 
-let load_music name ?(intro = 0.) ?(loop = Float.max_float) areas music_volume =
+let load_music dir name extension ?(intro = 0.) ?(loop = Float.max_float) music_volume =
   let music =
-    Raylib.load_music_stream (File.make_assets_path [ "audio"; "music"; fmt "%s.ogg" name ])
+    Raylib.load_music_stream (File.make_assets_path [ "audio"; dir; fmt "%s.%s" name extension ])
   in
   Raylib.set_music_volume music music_volume;
-  (* TODO this isn't a good way to do this
-     - this is starting all ~8 area musics, and then only advancing the one for the current area
-     - should probably be using stop_music_stream
-  *)
   Raylib.play_music_stream music;
-  { areas; music = { name; t = music; loop_start = { at = intro }; loop_end = { at = loop } } }
+  { name; t = music; loop_start = { at = intro }; loop_end = { at = loop } }
+
+let load_area_music name ?(intro = 0.) ?(loop = Float.max_float) areas music_volume : area_music =
+  let music = load_music "music" name "ogg" ~intro ~loop music_volume in
+  { areas; music }
+
+let load_continuous_sound ?(loop = Float.max_float) name music_volume : music =
+  load_music "continuous" name "wav" ~intro:0. ~loop music_volume
 
 let load_sound name =
   let path = File.make_assets_path [ "audio"; "sound-effects"; fmt "%s.wav" name ] in